# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for E-Document Management System",
        "title": "E-Document API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@edocument.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/api",
    "paths": {
...
        "/v1/outgoing-docs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all outgoing documents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outgoing-docs"
                ],
                "summary": "Get all outgoing documents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.OutgoingDocResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new outgoing document record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outgoing-docs"
                ],
                "summary": "Create outgoing document",
                "parameters": [
                    {
                        "description": "Outgoing Document data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_domain.CreateOutgoingDocRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_domain.OutgoingDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/outgoing-docs/department/{deptId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get outgoing documents for a specific department",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outgoing-docs"
                ],
                "summary": "Get outgoing documents by department",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department UUID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.OutgoingDocResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/outgoing-docs/user/{userId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get outgoing documents for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outgoing-docs"
                ],
                "summary": "Get outgoing documents by user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.OutgoingDocResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/outgoing-docs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get outgoing document details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outgoing-docs"
                ],
                "summary": "Get outgoing document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Outgoing Document ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_domain.OutgoingDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "e-document-backend_internal_domain.CreateOutgoingDocRequest": {
            "type": "object",
            "required": [
                "doc_id"
            ],
            "properties": {
                "doc_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "e-document-backend_internal_domain.IncomingDocResponse": {
            "type": "object",
            "properties": {
                "approver_date": {
                    "type": "string"
                },
                "approver_id": {
                    "type": "string"
                },
                "approver_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "dept_id": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "doc_name": {
                    "type": "string"
                },
                "doc_no": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "incoming_no": {
                    "type": "string"
                },
                "received_date": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "string"
                },
                "receiver_name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                },
                "sender_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocStatus"
                }
            }
        },
        "e-document-backend_internal_domain.IncomingDocStatus": {
            "type": "string",
            "enum": [
                "pending",
                "received",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "IncomingStatusPending",
                "IncomingStatusReceived",
                "IncomingStatusApproved",
                "IncomingStatusRejected"
            ]
        },
...
        "e-document-backend_internal_domain.OutgoingDocResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dept_id": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "doc_name": {
                    "type": "string"
                },
                "doc_no": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "incoming_docs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocResponse"
                    }
                },
                "outgoing_no": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "e-document-backend_internal_util.ErrorCode": {
            "type": "string",
            "enum": [
                "UNAUTHORIZED",
                "INCORRECT_PASSWORD",
                "INVALID_CREDENTIALS",
                "TOKEN_EXPIRED",
                "INVALID_TOKEN",
                "FORBIDDEN",
                "VALIDATION_ERROR",
                "MISSING_REQUIRED_FIELD",
                "INVALID_INPUT",
                "INTERNAL_SERVER_ERROR",
                "DATABASE_ERROR",
                "CONFIG_NOT_SET",
                "USER_NOT_FOUND",
                "USER_ALREADY_EXISTS",
                "EMAIL_ALREADY_EXISTS",
                "ROLE_NOT_FOUND",
                "ROLE_ALREADY_EXISTS"
            ],
            "x-enum-varnames": [
                "UNAUTHORIZED",
                "INCORRECT_PASSWORD",
                "INVALID_CREDENTIALS",
                "TOKEN_EXPIRED",
                "INVALID_TOKEN",
                "FORBIDDEN",
                "VALIDATION_ERROR",
                "MISSING_REQUIRED_FIELD",
                "INVALID_INPUT",
                "INTERNAL_SERVER_ERROR",
                "DATABASE_ERROR",
                "CONFIG_NOT_SET",
                "USER_NOT_FOUND",
                "USER_ALREADY_EXISTS",
                "EMAIL_ALREADY_EXISTS",
                "ROLE_NOT_FOUND",
                "ROLE_ALREADY_EXISTS"
            ]
        },
...
        "e-document-backend_internal_util.PaginationInfo": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "itemsPerPage": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "e-document-backend_internal_util.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "$ref": "#/definitions/e-document-backend_internal_util.ErrorCode"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/e-document-backend_internal_util.PaginationInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}